+++ apps/maimai-web/src/routes/api/jobs/data/+server.ts
@@ -47,6 +47,7 @@ export const POST: RequestHandler = async ({ request }) => {
     playerDataHtml,
     allMusicRecordHtml,
     imgGenInput,
+    fullPlayData,
     version,
   } = bodyResult.data;
 
@@ -196,6 +197,7 @@ export const POST: RequestHandler = async ({ request }) => {
       playerDataHtml,
       allMusicRecordHtml,
       dataForImageGen: JSON.stringify(imgGenInput),
+      fullPlayData: fullPlayData ? JSON.stringify(fullPlayData) : null,
     });
 
     try {